Move storage migration to migration file
